+++ Sources/RepoBar/Support/AppLimits.swift
@@ -31,6 +31,7 @@ enum AppLimits {
 
     enum GitHubReferenceMonitor {
         static let minimumBareDigits = 1
+        static let queryLimit = 8
         static let cacheLookupLimit = 20
         static let liveLookupLimit = 80
     }